Reafactored to make a more generic and reusable GifListView component
