+++ src/reducers/initialState.js
@@ -1,3 +1,5 @@
 export default {
+  searchQuery: '',
+  searchResults: [],
   trendingGifs: []
 };
